Modifiche ad AIComponent, sistemare update worldX e worldY

+++ app/src/main/java/com/MattiaDiMeglio/progettogamedesign/GameObject.java
@@ -28,7 +28,7 @@ public class GameObject {
 
     public void updatePosition(int x, int y){};
 
-    public void updatePosition(int x, int y, int angle){};
+    public void updatePosition(float x, float y, int angle){};
 
     //public void updatePosition(int x, int y, int angle, int strength){};
